% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting.R
\name{pgrouting_hastighet}
\alias{pgrouting_hastighet}
\alias{pgrouting_hastighet_gang}
\alias{pgrouting_hastighet_cykel}
\alias{pgrouting_hastighet_elcykel}
\title{Standardhastigheter (km/h) för gång, cykel och elcykel}
\usage{
pgrouting_hastighet_gang()

pgrouting_hastighet_cykel()

pgrouting_hastighet_elcykel()
}
\value{
Ett tal.
}
\description{
Standardhastigheter (km/h) för gång, cykel och elcykel
}
