% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting.R
\name{pgrouting_installera_i_postgis_db}
\alias{pgrouting_installera_i_postgis_db}
\title{Installera pgRouting-tillägget i en PostGIS-databas}
\usage{
pgrouting_installera_i_postgis_db(con)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Installera pgRouting-tillägget i en PostGIS-databas
}
