% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting.R
\name{pgrouting_kostnadskolumner_transporttyp_graf}
\alias{pgrouting_kostnadskolumner_transporttyp_graf}
\title{Beräkna kostnadskolumner per transportsätt i en graf-tabell}
\usage{
pgrouting_kostnadskolumner_transporttyp_graf(
  con = "ruttanalyser",
  schema_natverk = "grafer",
  tabell_natverk = "nvdb_alla_adresser",
  kostnadskolumn_bil_f = "hastighetsgrans_f",
  kostnadskolumn_bil_b = "hastighetsgrans_b",
  kostnadskolumn_meter = "kostnad_meter",
  berakna_kostnad_bil = TRUE,
  berakna_kostnad_gang = TRUE,
  berakna_kostnad_cykel = TRUE,
  berakna_kostnad_elcykel = TRUE
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller databasnamn.}

\item{schema_natverk, tabell_natverk}{Graf-tabellen.}

\item{kostnadskolumn_bil_f, kostnadskolumn_bil_b}{Hastighetskolumner (km/h).}

\item{kostnadskolumn_meter}{Kolumn med längd i meter.}

\item{berakna_kostnad_bil, berakna_kostnad_gang, berakna_kostnad_cykel, berakna_kostnad_elcykel}{Slå på/av respektive beräkning.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Lägger till \code{kostnad_gang_min}, \code{kostnad_cykel_min}, \code{kostnad_elcykel_min},
\code{kostnad_bil_f_min}, \code{kostnad_bil_b_min} utifrån \code{kostnad_meter} och
hastighetsgränser.
}
