% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting.R
\name{pgrouting_punkttabell_koppla_till_pgr_graf}
\alias{pgrouting_punkttabell_koppla_till_pgr_graf}
\title{Koppla en punkttabell till en pgRouting-graf (närmaste nod-id)}
\usage{
pgrouting_punkttabell_koppla_till_pgr_graf(
  con = "ruttanalyser",
  schema_punkter = "punktlager",
  tabell_punkter = "adresser",
  geom_kol_punkter = "geom",
  id_kol_punkter = "gml_id",
  schema_natverk = "grafer",
  tabell_natverk = "nvdb_alla_adresser",
  generella_namn_prova = TRUE
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller databasnamn.}

\item{schema_punkter, tabell_punkter, geom_kol_punkter, id_kol_punkter}{Punkttabellen.}

\item{schema_natverk, tabell_natverk}{Grafen (\verb{<tabell>_vertices_pgr} måste finnas).}

\item{generella_namn_prova}{Prova \code{"id"} resp. \code{"geom"}/\code{"geometry"} om de
angivna kolumnnamnen inte finns.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Lägger till kolumnen \verb{nid_<tabell_natverk>} med id:t på närmaste graf-nod.
}
