% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting_rutt.R
\name{pgrouting_skapa_ny_graf_nvdb_koppla_till_punkter}
\alias{pgrouting_skapa_ny_graf_nvdb_koppla_till_punkter}
\title{Bygg om hela adressgrafen vid ny NVDB-version}
\usage{
pgrouting_skapa_ny_graf_nvdb_koppla_till_punkter(
  con_geodb = "geodata",
  con_rutt = "ruttanalyser",
  natverkstyp = "bil",
  punkter_fran_schema = "adresser",
  punkter_fran_tabell = "dalarna",
  punkter_till_tabell = "adresser",
  punkter_till_schema = "punktlager"
)
}
\arguments{
\item{con_geodb, con_rutt}{\code{DBIConnection}-objekt eller databasnamn.}

\item{natverkstyp}{\code{"bil"} eller \code{"alla"}.}

\item{punkter_fran_schema, punkter_fran_tabell}{Adresskälla i geodata.}

\item{punkter_till_schema, punkter_till_tabell}{Adressmål i ruttanalyser.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Orkestrerar hela kedjan: klipp nätverket, kopiera adresser till
ruttanalyser, hitta närmaste nod, bygg graf, beräkna kostnader, flytta
underlagstabeller och koppla alla punktlager till alla grafer.
}
