% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting.R
\name{pgrouting_tabell_till_pgrgraf}
\alias{pgrouting_tabell_till_pgrgraf}
\title{Bygg en pgRouting-topologi (source/target) på en linjetabell}
\usage{
pgrouting_tabell_till_pgrgraf(
  con = "ruttanalyser",
  schema_graf = "grafer",
  tabell_graf = "nvdb_alla_adresser",
  id_kol_graf = "rad_id",
  geom_kol_graf = "geom",
  tolerans = 0.001
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller databasnamn.}

\item{schema_graf, tabell_graf, id_kol_graf, geom_kol_graf}{Tabellen som blir graf.}

\item{tolerans}{Anslutningstolerans för \code{pgr_createTopology()}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Bygg en pgRouting-topologi (source/target) på en linjetabell
}
