% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgis.R
\name{postgis_aktivera_i_postgres_db}
\alias{postgis_aktivera_i_postgres_db}
\alias{postgis_installera_i_postgres_db}
\title{Aktivera PostGIS-tillägget i en databas}
\usage{
postgis_aktivera_i_postgres_db(con)

postgis_installera_i_postgres_db(con)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Aktivera PostGIS-tillägget i en databas
}
