% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgis.R
\name{postgis_flytta_tabell}
\alias{postgis_flytta_tabell}
\title{Flytta en tabell från ett schema till ett annat}
\usage{
postgis_flytta_tabell(con = "default", schema_fran, tabell_fran, schema_till)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{schema_fran, tabell_fran}{Källa.}

\item{schema_till}{Målschema.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Flytta en tabell från ett schema till ett annat
}
