% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isokroner.R
\name{postgis_isokroner_dela_upp_polygoner}
\alias{postgis_isokroner_dela_upp_polygoner}
\title{Klipp isokronpolygoner så att de inte överlappar}
\usage{
postgis_isokroner_dela_upp_polygoner(
  skickad_sf,
  kategori_kol,
  isoid_kol,
  geom_kol
)
}
\arguments{
\item{skickad_sf}{sf-objekt med isokroner.}

\item{kategori_kol}{Kolumn med kostnadsintervall/nivå.}

\item{isoid_kol}{Kolumn med punkt-id (om flera punkter).}

\item{geom_kol}{Geometrikolumnens namn.}
}
\value{
Ett sf-objekt utan överlapp mellan nivåer.
}
\description{
Klipp isokronpolygoner så att de inte överlappar
}
