% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrouting_rutt.R
\name{postgis_kopiera_punkttabell_koppla_till_pgr_graf}
\alias{postgis_kopiera_punkttabell_koppla_till_pgr_graf}
\title{Kopiera en punkttabell mellan databaser och koppla den till alla grafer}
\usage{
postgis_kopiera_punkttabell_koppla_till_pgr_graf(
  con_fran_databas = "geodata",
  con_till_databas = "ruttanalyser",
  schema_fran,
  tabell_fran,
  schema_till,
  tabell_till,
  geom_kol_punkter = "geom",
  id_kol_punkter = "id",
  schema_natverk = "grafer"
)
}
\arguments{
\item{con_fran_databas, con_till_databas}{\code{DBIConnection}-objekt eller
databasnamn (standard: geodata resp. ruttanalyser).}

\item{schema_fran, tabell_fran}{Källa.}

\item{schema_till, tabell_till}{Mål.}

\item{geom_kol_punkter, id_kol_punkter}{Kolumner i punkttabellen.}

\item{schema_natverk}{Schema där graferna ligger.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Kopiera en punkttabell mellan databaser och koppla den till alla grafer
}
