% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgis.R
\name{postgis_kopiera_tabell}
\alias{postgis_kopiera_tabell}
\title{Kopiera en tabell inom en PostGIS-databas}
\usage{
postgis_kopiera_tabell(
  con = "default",
  schema_fran,
  tabell_fran,
  schema_till,
  tabell_till,
  skriv_over = FALSE
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{schema_fran, tabell_fran}{Källa.}

\item{schema_till, tabell_till}{Mål.}

\item{skriv_over}{Skriv över måltabellen om den finns.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Kopiera en tabell inom en PostGIS-databas
}
