% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdgeorouting-package.R
\docType{package}
\name{rdgeorouting-package}
\alias{rdgeorouting}
\alias{rdgeorouting-package}
\title{rdgeorouting: PostGIS- och pgRouting-verktyg för Region Dalarnas Ruttanalyser}
\description{
Bygga och köra pgRouting-grafer av vägnätet (NVDB), skapa isokroner, beräkna restid/avstånd mellan från- och till-punkter, samt pendlingsanalyser (pendlingsnätverk, kraftfält, in-/utpendling på ruta). Kräver en PostGIS-databas med pgRouting. Utbrutet ur postgis-, pgrouting- och pendlingsavsnitten i func_GIS.R i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
